prebeta3... mass commit
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@25 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cv/cvcontours.cpp
@@ -1170,7 +1170,7 @@ icvFindContoursInInterval( const CvArr* src,
     if( !result )
         CV_ERROR( CV_StsNullPtr, "NULL double CvSeq pointer" );
 
-    if( contourHeaderSize < sizeof(CvContour))
+    if( contourHeaderSize < (int)sizeof(CvContour))
         CV_ERROR( CV_StsBadSize, "Contour header size must be >= sizeof(CvContour)" );
 
     CV_CALL( storage00 = cvCreateChildMemStorage(storage));
